--- pyproject.toml.old
@@ -1,3 +1,0 @@
-[build-system]
-requires = ["setuptools>=42"]
-build-backend = "setuptools.build_meta"
