+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = Beetstream
+name = BeetstreamNext
 version = 1.4.0
 author = Binary Brain
 author_email = me@sachabron.ch
